+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- See more button for SKUSelector with a lot of items in a variation.
+- Schema for `SKUSelector` wtih `seeMoreLabel, maxItems` prop.
 
 ## [3.43.5] - 2019-06-10
 ### Fixed